refactor: extract PullUpIndicatorIos component
